move parser pkg to gsorm pkg
