--- internal/parser/error.go
@@ -1,9 +1,0 @@
-package parser
-
-import "github.com/morikuni/failure"
-
-const (
-	errInvalidType   failure.StringCode = "InvalidType"
-	errInvalidValue  failure.StringCode = "InvalidValue"
-	errInvalidSyntax failure.StringCode = "InvalidSyntax"
-)
