--- internal/parser/update_model.go
@@ -1,103 +1,0 @@
-package parser
-
-import (
-	"fmt"
-	"reflect"
-
-	"github.com/champon1020/gsorm/internal"
-	"github.com/morikuni/failure"
-)
-
-// UpdateModelParser is the model parser for update statement.
-type UpdateModelParser struct {
-	model       reflect.Value
-	modelType   reflect.Type
-	Cols        []string
-	ColumnField map[int]int
-}
-
-// NewUpdateModelParser creates UpdateModelParser instance.
-func NewUpdateModelParser(cols []string, model interface{}) (*UpdateModelParser, error) {
-	mTyp := reflect.TypeOf(model)
-	if mTyp.Kind() != reflect.Ptr {
-		err := failure.New(errInvalidValue, failure.Message("model must be a pointer"))
-		return nil, err
-	}
-
-	parser := &UpdateModelParser{
-		model:     reflect.ValueOf(model).Elem(),
-		modelType: mTyp.Elem(),
-		Cols:      cols,
-	}
-	return parser, nil
-}
-
-// Parse converts model to SQL.
-func (p *UpdateModelParser) Parse() (*internal.SQL, error) {
-	var sql internal.SQL
-
-	switch p.modelType.Kind() {
-	case reflect.Struct:
-		p.ParseStruct(&sql, p.model)
-		return &sql, nil
-	case reflect.Map:
-		if err := p.ParseMap(&sql, p.model); err != nil {
-			return nil, err
-		}
-		return &sql, nil
-	}
-
-	err := failure.New(errInvalidType,
-		failure.Context{"type": p.modelType.Kind().String()},
-		failure.Message("invalid type for internal.InsertModelParser.Parse"))
-	return nil, err
-}
-
-// ParseMap parses map to SQL.
-func (p *UpdateModelParser) ParseMap(sql *internal.SQL, model reflect.Value) error {
-	for i, c := range p.Cols {
-		if i > 0 {
-			sql.Write(",")
-		}
-		v := model.MapIndex(reflect.ValueOf(c))
-		if !v.IsValid() {
-			return failure.New(errInvalidSyntax,
-				failure.Message("column names must be included in one of map keys"))
-		}
-		s := internal.ToString(v.Interface(), nil)
-		sql.Write(fmt.Sprintf("%s = %s", c, s))
-	}
-	return nil
-}
-
-// ParseStruct parses struct to SQL.
-func (p *UpdateModelParser) ParseStruct(sql *internal.SQL, model reflect.Value) {
-	if p.ColumnField == nil {
-		p.ColumnField = p.columnsAndFields(model.Type())
-	}
-	for i := 0; i < len(p.Cols); i++ {
-		if i > 0 {
-			sql.Write(",")
-		}
-		opt := &internal.ToStringOpt{Quotes: true}
-		s := internal.ToString(model.Field(p.ColumnField[i]).Interface(), opt)
-		sql.Write(fmt.Sprintf("%s = %s", p.Cols[i], s))
-	}
-}
-
-func (p *UpdateModelParser) columnsAndFields(target reflect.Type) map[int]int {
-	cf := make(map[int]int)
-	for i, col := range p.Cols {
-		for j := 0; j < target.NumField(); j++ {
-			c := internal.ExtractTag(target.Field(j)).Column
-			if c == "" {
-				c = internal.SnakeCase(target.Field(j).Name)
-			}
-			if col != c {
-				continue
-			}
-			cf[i] = j
-		}
-	}
-	return cf
-}
